+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.2-dev
  - remove unnecessary `GooseAttack.number_of_cpus` instead calling `num_cpus::get()` directly
+ - remove `tests/gaggle.rs`, instead mixing gaggle tests with per-feature integration tests
 
 ## 0.10.1 Sep 20, 2020
  - rework `hatch_rate` to be stored in an `Option<usize>` as it can be `None` on a Worker